+++ fotoobo/cli/ems/get_commands.py
@@ -9,7 +9,7 @@ from fotoobo.helpers import cli_path
 from fotoobo.helpers.output import print_datatable
 from fotoobo.utils import ems
 
-app = typer.Typer()
+app = typer.Typer(no_args_is_help=True)
 log = logging.getLogger("fotoobo")