+++ fotoobo/cli/fmg/get_commands.py
@@ -9,7 +9,7 @@ from fotoobo.helpers import cli_path
 from fotoobo.helpers.output import print_datatable, write_policy_to_html
 from fotoobo.utils import fmg
 
-app = typer.Typer()
+app = typer.Typer(no_args_is_help=True)
 log = logging.getLogger("fotoobo")