+++ fotoobo/cli/fmg/main.py
@@ -9,7 +9,7 @@ from fotoobo.cli.fmg import get_commands as get
 from fotoobo.helpers import cli_path
 from fotoobo.utils import fmg
 
-app = typer.Typer()
+app = typer.Typer(no_args_is_help=True)
 log = logging.getLogger("fotoobo")