+++ fotoobo/cli/fmg/main.py
@@ -9,7 +9,7 @@ from fotoobo.cli.fmg import get_commands as get
 from fotoobo.helpers import cli_path
 from fotoobo.utils import fmg
 
-app = typer.Typer()
+app = typer.Typer(no_args_is_help=True, rich_markup_mode="rich")
 log = logging.getLogger("fotoobo")
 
 
@@ -25,48 +25,56 @@ def callback(context: typer.Context) -> None:
     log.debug("about to execute command: '%s'", context.invoked_subcommand)
 
 
-@app.command()
+@app.command(no_args_is_help=True)
 def assign(
-    host: str = typer.Argument(
-        "fmg",
-        help="The FortiManager hostname to access (must be defined in inventory)",
-        metavar="[host]",
-    ),
     adoms: str = typer.Argument(
         ...,
-        help="The ADOMs to assign the global policy/objects. Use 'fmg get adoms' to get a list of "
-        + "available ADOMs. Separate multiple ADOMs by comma (no spaces)",
+        help="The ADOMs to assign the global policy/objects to. Use "
+        "'fotoobo fmg get adoms' to get a list of "
+        "available ADOMs. Separate multiple ADOMs by comma (no spaces).",
         metavar="[adoms]",
         show_default=False,
+    ),
+    host: str = typer.Argument(
+        "fmg",
+        help="The FortiManager to access (must be defined in the inventory).",
+        metavar="[host]",
     ),
     timeout: int = typer.Option(
         60,
         "--timeout",
         "-t",
-        help="The timeout to wait for the FortiManager task to finish",
+        help="The timeout to wait for the FortiManager task to finish.",
         metavar="[timeout]",
     ),
 ) -> None:
     """
-    Assign a global policy to a specified ADOM or to a list of ADOMs
+    Assign a global policy to a specified ADOM or to a list of ADOMs.
     """
     fmg.assign(host, adoms, timeout=timeout)
 
 
-@app.command()
-def set(  # pylint: disable=redefined-builtin
+@app.command(no_args_is_help=True)
+def post(
     host: str = typer.Argument(
         ...,
-        help="The FortiManager hostname to access (must be defined in inventory)",
+        help="The FortiManager to access (must be defined in the inventory).",
         metavar="[host]",
+        show_default=False,
     ),
     file: str = typer.Argument(
-        ..., help="json file with payload(s)", show_default=False, metavar="[file]"
+        ..., help="JSON file with payload(s).", show_default=False, metavar="[file]"
     ),
-    adom: str = typer.Argument(..., help="The ADOM to issue the set command(s)", metavar="[adom]"),
+    adom: str = typer.Argument(
+        ..., help="The ADOM to issue the set command(s).", metavar="[adom]", show_default=False
+    ),
 ) -> None:
-    """FortiManager set command"""
-    fmg.set(host, file, adom)
+    """
+    POST any valid JSON request to the FortiManager.
+
+    Configure the FortiManager with any valid API call(s) given within the JSON file.
+    """
+    fmg.post(host, file, adom)
 
 
-app.add_typer(get.app, name="get", help="FortiManager get commands")
+app.add_typer(get.app, name="get", help="FortiManager get commands.")