+++ tests/cli/fmg/test_cli_fmg_main.py
@@ -17,7 +17,7 @@ def test_cli_app_fmg_help() -> None:
     arguments, options, commands = parse_help_output(result.stdout)
     assert not arguments
     assert options == {"-h", "--help"}
-    assert set(commands) == {"assign", "get", "set"}
+    assert set(commands) == {"assign", "get", "post"}
 
 
 def test_cli_app_fmg_assign_help() -> None:
@@ -32,7 +32,7 @@ def test_cli_app_fmg_assign_help() -> None:
 
 def test_cli_app_fmg_set_help() -> None:
     """Test cli help for fmg set"""
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fmg", "set", "-h"])
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fmg", "post", "-h"])
     assert result.exit_code == 0
     arguments, options, commands = parse_help_output(result.stdout)
     assert set(arguments) == {"host", "file", "adom"}