+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 All notable changes to this project will be documented in this file.
 For examples and guidelines see [https://keepachangelog.com/](https://keepachangelog.com/)
 
-## [Unreleased]
+# [Unreleased]
 
 
 
-## [Released]
+# [Released]
 
 ## [0.8.0] - 2023-04-13